+++ cvat-ui/src/components/annotation-page/top-bar/right-group.tsx
@@ -46,7 +46,7 @@ function RightGroup(props: Props): JSX.Element {
     const openGuide = useCallback(() => {
         const PADDING = Math.min(window.screen.availHeight, window.screen.availWidth) * 0.4;
         jobInstance.guide().then((guide) => {
-            if (guide?.markdown) {
+            if (guide) {
                 Modal.info({
                     icon: null,
                     width: window.screen.availWidth - PADDING,
