+++ cvat-ui/src/components/md-guide/annotation-guide-page.tsx
